+++ src/components/ui/drawer/index.tsx
@@ -1,12 +1,11 @@
-import { css } from '@styled-system/css'
-import { AnimatePresence, motion, PanInfo } from 'framer-motion'
-import { ReactNode, useCallback, useMemo } from 'react'
+import { AnimatePresence } from 'framer-motion'
+import { ReactNode } from 'react'
 
 import BackLayer from '@/components/ui/drawer/BackLayer'
+import DrawerBody from '@/components/ui/drawer/DrawerBody'
 
 interface DrawerProps {
   isOpen: boolean
-  openHeight: number
   close: () => void
   children: ReactNode
 }
@@ -16,59 +15,16 @@ interface DrawerProps {
  * 이미 Timetable에서 사용하는 고유한 UI 이름을 BottomSheet으로 명명하여
  * Drawer로 사용
  *
- * useDrawer hook에서 제공하는 isOpen과 close 이용,
- * openHeight는 열릴 때의 높이를 의미
  */
-const Drawer = ({ isOpen, openHeight, close, children }: DrawerProps) => {
-  const expandedHeight = useMemo(() => Math.min(openHeight, window.innerHeight), [openHeight])
-
-  const onDragEnd = useCallback(
-    (_: MouseEvent | TouchEvent | PointerEvent, info: PanInfo) => {
-      const offsetThreshold = 150
-      const deltaThreshold = 5
-
-      const isOverOffsetThreshold = Math.abs(info.offset.y) > offsetThreshold
-      const isOverDeltaThreshold = Math.abs(info.delta.y) > deltaThreshold
-
-      const isOverThreshold = isOverOffsetThreshold || isOverDeltaThreshold
-
-      if (!isOverThreshold) return
-
-      info.offset.y >= 0 && close()
-    },
-    [close],
-  )
-
+const Drawer = ({ isOpen, close, children }: DrawerProps) => {
   return (
     <AnimatePresence>
-      {isOpen && <BackLayer close={close} />}
-      <motion.div
-        key="drawer_body"
-        className={css({
-          pos: 'fixed',
-          top: '100dvh',
-          left: 0,
-          w: 'full',
-          h: '100dvh',
-          bgColor: 'white',
-          rounded: '20px 20px 0 0',
-          willChange: 'transform',
-          pt: '15px',
-          zIndex: 101,
-          display: 'flex',
-          flexDir: 'column',
-        })}
-        drag={'y'}
-        dragConstraints={{ top: 0, bottom: 0 }}
-        dragElastic={0.4}
-        animate={{ top: isOpen ? `calc(100dvh - ${expandedHeight}px)` : '100dvh' }}
-        onDragEnd={onDragEnd}
-      >
-        <div
-          className={css({ bgColor: 'darkGray.2', rounded: 'full', h: 1, w: '49px', alignSelf: 'center', mb: '15px' })}
-        />
-        {children}
-      </motion.div>
+      {isOpen && (
+        <>
+          <BackLayer close={close} />
+          <DrawerBody close={close}>{children}</DrawerBody>
+        </>
+      )}
     </AnimatePresence>
   )
 }
